Migrate TestAuthorization_Custom into a subfolder as a new suite because of custom control plane setup
