+++ tests/integration/security/main_test.go
@@ -16,12 +16,10 @@
 package security
 
 import (
-	"fmt"
 	"testing"
 
 	"istio.io/istio/pkg/test/framework"
 	"istio.io/istio/pkg/test/framework/components/istio"
-	"istio.io/istio/pkg/test/framework/components/namespace"
 	"istio.io/istio/pkg/test/framework/resource"
 	"istio.io/istio/tests/integration/security/util"
 )
@@ -45,46 +43,12 @@ func setupConfig(ctx resource.Context, cfg *istio.Config) {
 	if cfg == nil {
 		return
 	}
-
-	// Create the namespace instance ahead of time so that it can be used in the mesh config.
-	extAuthzServiceNamespace, extAuthzServiceNamespaceErr = namespace.New(ctx, namespace.Config{
-		Prefix: "test-ns-ext-authz-service",
-		Inject: true,
-	})
-	var extAuthzNamespace string
-	if extAuthzServiceNamespaceErr == nil {
-		extAuthzNamespace = extAuthzServiceNamespace.Name()
-	}
-	service := fmt.Sprintf("ext-authz.%s.svc.cluster.local", extAuthzNamespace)
-	serviceWithNamespace := fmt.Sprintf("%s/%s", extAuthzNamespace, service)
-
-	cfg.ControlPlaneValues = fmt.Sprintf(`
+	cfg.ControlPlaneValues = `
 meshConfig:
   accessLogEncoding: JSON
   accessLogFile: /dev/stdout
   defaultConfig:
     gatewayTopology:
       numTrustedProxies: 1
-  extensionProviders:
-  - name: "ext-authz-http"
-    envoyExtAuthzHttp:
-      service: %q
-      port: 8000
-      pathPrefix: "/check"
-      includeHeadersInCheck: ["x-ext-authz"]
-  - name: "ext-authz-grpc"
-    envoyExtAuthzGrpc:
-      service: %q
-      port: 9000
-  - name: "ext-authz-http-local"
-    envoyExtAuthzHttp:
-      service: ext-authz-http.local
-      port: 8000
-      pathPrefix: "/check"
-      includeHeadersInCheck: ["x-ext-authz"]
-  - name: "ext-authz-grpc-local"
-    envoyExtAuthzGrpc:
-      service: ext-authz-grpc.local
-      port: 9000
-`, service, serviceWithNamespace)
+`
 }
